simulator

+++ PL/ViewModel/Converters/DroneConverter.cs
@@ -69,7 +69,7 @@ namespace PL
                 Id = drone.DroneId,
                 Model = drone.DroneModel,
                 Battery = (int)drone.BatteryStatus,
-                Status = (Enums.DroneStatuses)drone.DroneStatus,
+                Status = (Enums.DroneStatuses)((int)drone.DroneStatus),
                 Weight = (Enums.WeightCategories)drone.Weight,
                 Location =LocationConverter.ConvertLocation(drone.DroneLocation),
                 DeliveryByTransfer = drone.DeliveryTransfer == null ? null : ParcelConverter.ConvertParcelInTransfer(drone.DeliveryTransfer)
